refactor: CouponCard가 유효한지 enable 여부를 계산해 props로 전달하도록 변경
- 대신 enable 여부를 계산해 props로 전달
-  CouponCard가 유효한지 확인하기 위해 배송비를 props로 받지 않도록 변경
